refactor : addSection을 LineService에서 SectionService로 변경
