+++ src/main/java/subway/controller/LineController.java
@@ -89,7 +89,7 @@ public class LineController {
 
         lineService.checkDuplicateName(lineId, upStationId, downStationId);
 
-        lineService.addSectionOnLine(line, section);
+        sectionService.addSection(line, section);
         SectionResponse response = new SectionResponse(section);
         return ResponseEntity.ok().body(response);
     }
